+++ src/test/ui/transmute/transmute-type-parameters.rs
@@ -13,18 +13,18 @@
 use std::mem::transmute;
 
 unsafe fn f<T>(x: T) {
-    let _: isize = transmute(x);
-//~^ ERROR differently sized types: T (size can vary) to isize
+    let _: i32 = transmute(x);
+//~^ ERROR differently sized types: T (size can vary) to i32
 }
 
-unsafe fn g<T>(x: (T, isize)) {
-    let _: isize = transmute(x);
-//~^ ERROR differently sized types: (T, isize) (size can vary because of T) to isize
+unsafe fn g<T>(x: (T, i32)) {
+    let _: i32 = transmute(x);
+//~^ ERROR differently sized types: (T, i32) (size can vary because of T) to i32
 }
 
 unsafe fn h<T>(x: [T; 10]) {
-    let _: isize = transmute(x);
-//~^ ERROR differently sized types: [T; 10] (size can vary because of T) to isize
+    let _: i32 = transmute(x);
+//~^ ERROR differently sized types: [T; 10] (size can vary because of T) to i32
 }
 
 struct Bad<T> {
@@ -32,8 +32,8 @@ struct Bad<T> {
 }
 
 unsafe fn i<T>(x: Bad<T>) {
-    let _: isize = transmute(x);
-//~^ ERROR differently sized types: Bad<T> (size can vary because of T) to isize
+    let _: i32 = transmute(x);
+//~^ ERROR differently sized types: Bad<T> (size can vary because of T) to i32
 }
 
 enum Worse<T> {
@@ -42,13 +42,13 @@ enum Worse<T> {
 }
 
 unsafe fn j<T>(x: Worse<T>) {
-    let _: isize = transmute(x);
-//~^ ERROR differently sized types: Worse<T> (size can vary because of T) to isize
+    let _: i32 = transmute(x);
+//~^ ERROR differently sized types: Worse<T> (size can vary because of T) to i32
 }
 
 unsafe fn k<T>(x: Option<T>) {
-    let _: isize = transmute(x);
-//~^ ERROR differently sized types: std::option::Option<T> (size can vary because of T) to isize
+    let _: i32 = transmute(x);
+//~^ ERROR differently sized types: std::option::Option<T> (size can vary because of T) to i32
 }
 
 fn main() {}
